Move postcode parsing to auth schemes
